Move wait_for_app into itest utils so we never use it in production code
