+++ paasta_itests/steps/paasta_metastatus_steps.py
@@ -13,11 +13,11 @@
 # limitations under the License.
 import os
 
+import itest_utils
 from behave import then
 from behave import when
 from marathon import MarathonApp
 
-from paasta_tools import marathon_tools
 from paasta_tools.utils import _run
 from paasta_tools.utils import remove_ansi_escape_sequences
 
@@ -74,7 +74,7 @@ def marathon_task_is_ready(context, app_id):
 @when(u'{num:d} tasks belonging to the app with id "{app_id}" are in the task list')
 def marathon_tasks_are_ready(context, num, app_id):
     """Wait for the specified number of  tasks with matching task names to be ready. time out in 60 seconds """
-    marathon_tools.wait_for_app_to_launch_tasks(context.marathon_client, app_id, num)
+    itest_utils.wait_for_app_to_launch_tasks(context.marathon_client, app_id, num)
 
 
 @then(u'paasta_metastatus{flags} exits with return code "{expected_return_code}" and output "{expected_output}"')
